pyproject.toml
[tool.poetry]
name = "ai-assistant"
version = "0.0.1"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
Flask = "^3.0.3"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
openai = "^1.51.0"
PyYAML = "^6.0.2"
pandas = "^2.2.3"
neo4j = "^4.4"
biocypher = "^0.6.1"
scikit-learn = "^1.5.2"
qdrant-client = "^1.12.0"
autogen = "^0.3.1"
tiktoken = "^0.8.0"
google-generativeai = "^0.8.3"
pyjwt = "^2.10.0"
flask-limiter = "^3.8.0"
gunicorn = "^23.0.0"
pypdf2 = "^3.0.1"
pytest = "^8.3.4"
autogen-agentchat = {version = ">=0.2,<1.0", extras = ["gemini", "llm"]}
flask-cors = "^5.0.0"
flask-socketio = "^5.5.1"
eventlet = "^0.39.1"
redis = "^6.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
